split into ffi
