+++ .gitignore
@@ -10,6 +10,7 @@ target
 
 # OS junk
 **/.DS_Store
+**/._.DS_Store
 
 # Ctags file
 tags
