+++ boot/start.S
@@ -1,15 +1,16 @@
 .text
 .global _start
 
+#define arg1 %rdi
+
+# Xen places the address of the start_info struct into the SI register.
+#define start_info_addr %rsi
+
 _start:
     mov $STACK_TOP, %rbp
     mov %rbp, %rsp
     # Xen places the address of the start_info struct into the SI register.
-    mov %rsi, start_info_page
+    mov start_info_addr, arg1
     call prologue
     call poweroff
 .bss
-
-.global start_info_page
-start_info_page:
-    .quad 0
