+++ src/main.rs
@@ -24,11 +24,12 @@ mod std;
 mod xen;
 
 pub use xen::poweroff;
-pub use xen::console_io::STDOUT;
+pub use xen::STDOUT;
 pub use xen::mem::sbrk;
-pub use xen::emergency_console::EMERGENCY_CONSOLE as DEBUG;
-use xen::start_info::start_info_page;
+pub use xen::DEBUG;
+use xen::StartInfoPage;
 use core::fmt::Write;
+use core::ptr;
 use alloc::boxed::Box;
 
 #[lang = "eh_personality"]
@@ -44,26 +45,15 @@ pub extern fn rust_begin_unwind(args: core::fmt::Arguments, file: &'static str, 
     xen::crash();
 }
 
-fn print_init_info(){
-    writeln!(STDOUT, "Magic: {}", core::str::from_utf8(&start_info_page.magic).unwrap_or("ERROR")).unwrap();
-    writeln!(STDOUT, "nr_pages: {:#X}", start_info_page.nr_pages).unwrap();
-    writeln!(STDOUT, "shared_info: {:#X}", start_info_page.shared_info).unwrap();
-}
-
-
 #[no_mangle]
-pub extern fn prologue() {
+pub extern fn prologue(start_info_page : *const StartInfoPage) {
     unsafe {
-        writeln!(DEBUG, "prologue!").unwrap();
+        writeln!(DEBUG, "start_info_page at {}, {:o}", 1234, start_info_page as usize).unwrap();
+        let page = ptr::read(start_info_page);
+        xen::initialize(page);
         use core::ptr;
         let null: *const u8 = ptr::null();
         let argv: *const *const u8 = &null;
-        writeln!(DEBUG, "mm::setup").unwrap();
-        writeln!(DEBUG, "xen::console_io::initialize").unwrap();
-        xen::console_io::initialize();
-        writeln!(DEBUG, "xen::xenstore::initialize").unwrap();
-        xen::xenstore::initialize();
-        writeln!(DEBUG, "end of prologue!").unwrap();
         let _result = main(0, argv);
     }
 }
@@ -110,7 +100,6 @@ pub fn main(_argc: isize, _argv: *const *const u8) -> isize {
     }
 
     writeln!(DEBUG, "done!").unwrap();
-    print_init_info();
 
     0
 }
