+++ src/xen/console.rs
@@ -1,46 +1,35 @@
 use core::mem::size_of_val;
 use std::io;
+use std::io::Write;
 use std::sync::RwLock;
 use xen::arch::mem;
-use ::xen::event_channels::*;
+use xen::event_channels::*;
+use xen::ffi::console::*;
 
-pub static CONSOLE: RwLock<Option<Console<'static>>> = RwLock::new(Option::None);
+pub static CONSOLE: RwLock<Option<Console>> = RwLock::new(Option::None);
 
-pub struct Console<'a> {
-    pub interface: &'a mut xencons_interface,
+pub struct Console {
+    pub interface: xencons_interface,
     pub event_channel: EventChannel
 }
 
-const INPUT_RING_SIZE: usize = 1024;
-const OUTPUT_RING_SIZE: usize = 2048;
-
-#[repr(C)]
-pub struct xencons_interface {
-    input       : [u8; INPUT_RING_SIZE],           // renamed because 'in' is a keyword
-    output      : [u8; OUTPUT_RING_SIZE],
-    in_cons     : u32,
-    in_prod     : u32,
-    out_cons    : u32,
-    out_prod    : u32
-}
-
-impl<'a> Console<'a> {
+impl Console {
     fn write_byte(&mut self, byte: u8) {
         while {
             let data = (self.interface.out_prod - self.interface.out_cons) as usize;
             unsafe { self.event_channel.notify(); }
             mem::mb();
-            data >= size_of_val(&self.interface.output)
+            data >= size_of_val(&self.interface.out)
         } {}
 
         let ring_index = mod_output_ring_size(self.interface.out_prod) as usize;
         
-        self.interface.output[ring_index] = byte;
+        self.interface.out[ring_index] = byte;
         self.interface.out_prod = self.interface.out_prod.wrapping_add(1);
     }
 }
 
-impl<'a> io::Write for Console<'a> {
+impl io::Write for Console {
     //Listing 6.4 in The Definitive Guide to the Xen Hypervisor
     fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
         for b in buf {
@@ -57,6 +46,10 @@ impl<'a> io::Write for Console<'a> {
     }
 }
 
-fn mod_output_ring_size(i: u32) -> u32 {
-    i & ((OUTPUT_RING_SIZE as u32) - 1)
+pub unsafe fn initialize(interface: xencons_interface, event_channel: EventChannel) {
+    let console = CONSOLE.write();
+    *console = Some (Console {
+        event_channel: event_channel,
+        interface: interface
+    })
 }
