+++ src/xen/emergency_console.rs
@@ -1,9 +1,8 @@
 use std::sync::RwLock;
-use ::xen::hypercalls;
-use ::xen::hypercalls::console_io;
+use xen::ffi::hypercalls::console_io::write;
 use core::fmt;
 
-static _EMERGENCY_CONSOLE : RwLock<EmergencyConsoleInterface> = 
+static EMERGENCY_CONSOLE : RwLock<EmergencyConsoleInterface> = 
     RwLock::new(
         EmergencyConsoleInterface { 
             buffer: [0; LINE_LENGTH + 1]
@@ -17,36 +16,31 @@ struct EmergencyConsoleInterface {
 }
 
 #[allow(non_camel_case_types)]
-pub struct EMERGENCY_CONSOLE;
+pub struct DEBUG;
 
-impl EMERGENCY_CONSOLE {
-    fn notify(buf: &[u8], i: usize) {
-        hypercall!(i64, hypercalls::Command::console_io, console_io::SubCommand::write, i, buf.as_ptr());
-    }
-}
-
-impl fmt::Write for EMERGENCY_CONSOLE {
+impl fmt::Write for DEBUG {
     fn write_str(&mut self, s: &str) -> fmt::Result {
-        let mut buffer = _EMERGENCY_CONSOLE.write().buffer;
+        let mut buffer = EMERGENCY_CONSOLE.write().buffer;
         let mut i = 0;
         for c in s.as_bytes() {
             if *c == '\0' as u8 {
                 break;
             } else {
-                if i < LINE_LENGTH {
-                    buffer[i] = *c;
-                    i = i + 1;
-                } else {
-                    EMERGENCY_CONSOLE::notify(&buffer, LINE_LENGTH + 1);
+                if i == LINE_LENGTH {
+                    buffer[i] = '\0' as u8;
+                    write(&buffer);
                     i = 0;
                 }
+                buffer[i] = *c;
+                i = i + 1;
             }
         }
 
         if i != 0 {
             buffer[i] = '\0' as u8;
             i += 1;
-            EMERGENCY_CONSOLE::notify(&buffer, i);
+            let to_write = &buffer[..i];
+            write(to_write);
         }
         Ok(())
     }
