+++ src/xen/ffi/shared_info.rs
