--- src/xen/hypercalls/console_io/mod.rs
@@ -1,6 +1,0 @@
-#[repr(usize)]
-#[allow(non_camel_case_types)]
-pub enum SubCommand {
-    write = 0,
-    read  = 1
-}
