--- src/xen/hypercalls/event_channel_op/alloc_unbound.rs
@@ -1,12 +1,0 @@
-use ::xen::hypercalls::DomID;
-use ::xen::hypercalls::event_channel_op::Port;
-
-#[repr(C)]
-#[derive(Clone,Copy)]
-pub struct Args {
-    pub dom: DomID,
-    pub remote_dom: DomID,
-    /// Output
-    pub port: Port
-}
-
