--- src/xen/hypercalls/event_channel_op/bind_interdomain.rs
@@ -1,11 +1,0 @@
-use ::xen::hypercalls::DomID;
-use ::xen::hypercalls::event_channel_op::Port;
-
-#[repr(C)]
-#[derive(Clone,Copy)]
-pub struct Args {
-    pub remote_dom: DomID,
-    pub remote_port: Port,
-    /// Output
-    pub local_port: Port
-}
