--- src/xen/hypercalls/event_channel_op/close.rs
@@ -1,7 +1,0 @@
-use ::xen::hypercalls::event_channel_op::Port;
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-pub struct Args {
-    pub port: Port
-}
